Move the registrar features to a separate file
